requirements.txt
numpy~=1.19.5
pandas~=1.2.3
pyserial~=3.5
pyqtgraph~=0.11.1
PyQt5~=5.15.2

# 均使用pip install安装
urx~=0.11.0
fire~=0.4.0